Made connection in a diagram reversible
