+++ gaphor/diagram/diagramtools/placement.py
@@ -1,17 +1,13 @@
 import logging
 from typing import Callable, Optional, Type, TypeVar
 
-from gaphas.aspect.connector import Connector as ConnectorAspect
-from gaphas.aspect.connector import ItemConnector
 from gaphas.aspect.handlemove import HandleMove
 from gaphas.tool.itemtool import MoveType
 from gaphas.view import GtkView
 from gi.repository import Gtk
 
-from gaphor.core import transactional
 from gaphor.core.eventmanager import EventManager
 from gaphor.core.modeling import Diagram, Element
-from gaphor.diagram.connectors import Connector
 from gaphor.diagram.event import DiagramItemPlaced
 from gaphor.diagram.grouping import Group
 from gaphor.diagram.presentation import Presentation
@@ -131,98 +127,3 @@ def new_item_factory(
 
     item_factory.item_class = item_class  # type: ignore[attr-defined]
     return item_factory
-
-
-@ConnectorAspect.register(Presentation)
-class PresentationConnector(ItemConnector):
-    """Handle Tool (acts on item handles) that uses the Connector protocol to
-    connect items to one-another.
-
-    It also adds handles to lines when a line is grabbed on the middle
-    of a line segment (points are drawn by the LineSegmentPainter).
-    """
-
-    def allow(self, sink):
-        adapter = Connector(sink.item, self.item)
-        return adapter and adapter.allow(self.handle, sink.port)
-
-    @transactional
-    def connect(self, sink):
-        """Create connection at handle level and at model level."""
-        handle = self.handle
-        item = self.item
-        cinfo = self.connections.get_connection(handle)
-
-        try:
-            callback = DisconnectHandle(self.item, self.handle, self.connections)
-            if cinfo and cinfo.connected is sink.item:
-                # reconnect only constraint - leave model intact
-                log.debug("performing reconnect constraint")
-                constraint = sink.port.constraint(item, handle, sink.item)
-                self.connections.reconnect_item(
-                    item, handle, sink.port, constraint=constraint
-                )
-            elif cinfo:
-                # first disconnect but disable disconnection handle as
-                # reconnection is going to happen
-                adapter = Connector(sink.item, item)
-                try:
-                    connect = adapter.reconnect
-                except AttributeError:
-                    connect = adapter.connect
-                else:
-                    cinfo.callback.disable = True
-                self.disconnect()
-
-                # new connection
-                self.connect_handle(sink, callback=callback)
-
-                # adapter requires both ends to be connected.
-                connect(handle, sink.port)
-            else:
-                # new connection
-                adapter = Connector(sink.item, item)
-                self.connect_handle(sink, callback=callback)
-                adapter.connect(handle, sink.port)
-        except Exception:
-            log.error("Error during connect", exc_info=True)
-
-    @transactional
-    def disconnect(self):
-        super().disconnect()
-
-
-class DisconnectHandle:
-    """Callback for items disconnection using the adapters.
-
-    This is an object so disconnection data can be serialized/deserialized
-    using pickle.
-
-    :Variables:
-     item
-        Connecting item.
-     handle
-        Handle of connecting item.
-     disable
-        If set, then disconnection is disabled.
-    """
-
-    def __init__(self, item, handle, connections):
-        self.item = item
-        self.handle = handle
-        self.connections = connections
-        self.disable = False
-
-    def __call__(self):
-        handle = self.handle
-        item = self.item
-        connections = self.connections
-        cinfo = connections.get_connection(handle)
-
-        if self.disable:
-            log.debug(f"Not disconnecting {item}.{handle} (disabled)")
-        else:
-            log.debug(f"Disconnecting {item}.{handle}")
-            if cinfo:
-                adapter = Connector(cinfo.connected, item)
-                adapter.disconnect(handle)
